🏗️ refactor: JoinChallengeClient 컴포넌트 분리
