+++ src/app/join/page.tsx
@@ -1,4 +1,4 @@
-import JoinChallengeClient from "./JoinChallengeClient";
+import JoinChallengeClient from "@/components/join/JoinChallengeClient";
 
 export default function JoinChallengePage() {
   return <JoinChallengeClient />;
