+++ src/components/join/JoinChallengeClient.tsx
